IN PROGRESS - issue SOAPUI-3344: Implement sending initial request for getting locations https://jira.eviware.com/browse/SOAPUI-3344
Enabled sendProject() to make a real request to AlertSite. Still getting an error through.

The locations is fetch the first time a TestCase is opened and then cached for the remainder of the soapUI session.

The test is disabled until further.

Still need of better error handling and refactoring. See FIXME comments.