IN PROGRESS - issue SOAPUI-3345: Implement sending request to run a TestCase https://jira.eviware.com/browse/SOAPUI-3345
Moved Location, TestOnDemandCaller and TestOnDemandCallerTest to more suitiable packages.

Reduced redundancy in TestOnDemaindCaller.

Changed the behaviour of the endpoint system property to be the production endpoint as default and use a system property if defined to override the domain for the endpoint.

The TestOnDemandCallerTest is disabled until the soapui.testondemand.host is defined in Hudson.