IN PROGRESS - issue SOAPUI-3344: Implement sending initial request for getting locations https://jira.eviware.com/browse/SOAPUI-3344
Moving the call to the locations resource from start up to when opening the TestCase panel. The drop down will now also show an initializing message when the call has not been made and a message for when no locations has been found.

This is due to the soapUI Pro testrunner accually initializing StandaloneSoapUICore where the call was previously.